make tcp servers also cancelable
