+++ linkerd/app/integration/src/tcp.rs
@@ -9,6 +9,7 @@ use std::sync::Arc;
 use tokio::io::{AsyncReadExt, AsyncWriteExt};
 use tokio::net::TcpStream;
 use tokio::sync::{mpsc, oneshot};
+use tracing_futures::Instrument;
 
 type TcpSender = mpsc::UnboundedSender<oneshot::Sender<TcpConnSender>>;
 type TcpConnSender = mpsc::UnboundedSender<(
@@ -226,29 +227,40 @@ async fn run_server(tcp: TcpServer) -> server::Listening {
     let any_port = SocketAddr::from(([127, 0, 0, 1], 0));
     let std_listener = StdTcpListener::bind(&any_port).expect("bind");
     let addr = std_listener.local_addr().expect("local_addr");
-    let jh = tokio::spawn(async move {
-        let mut accepts = tcp.accepts;
+    let jh = tokio::spawn(
+        async move {
+            let mut accepts = tcp.accepts;
+            let (drain_tx, drain) = drain::channel();
+            let listen = async move {
+                let mut listener =
+                    TcpListener::from_std(std_listener).expect("TcpListener::from_std");
 
-        let listen = async move {
-            let mut listener = TcpListener::from_std(std_listener).expect("TcpListener::from_std");
+                loop {
+                    let (sock, _) = listener.accept().await.unwrap();
+                    let cb = accepts.pop_front().expect("no more accepts");
+                    srv_conn_count.fetch_add(1, Ordering::Release);
 
-            loop {
-                let (sock, _) = listener.accept().await.unwrap();
-                let cb = accepts.pop_front().expect("no more accepts");
-                srv_conn_count.fetch_add(1, Ordering::Release);
+                    let fut = cb.call_box(sock);
+                    tokio::spawn(cancelable(drain.clone(), async move {
+                        fut.await;
+                        Ok::<(), ()>(())
+                    }));
+                }
+            };
 
-                let fut = cb.call_box(sock);
-                tokio::task::spawn_local(fut);
+            let _ = started_tx.send(());
+            tokio::select! {
+                _ = rx => {
+                    tracing::trace!("shutting down...");
+                    drain_tx.drain().await;
+                    tracing::trace!("shut down");
+                },
+                _ = listen => { },
             }
-        };
-
-        let _ = started_tx.send(());
-        tokio::select! {
-            _ = rx => { },
-            _ = listen => { },
+            Ok(())
         }
-        Ok(())
-    });
+        .instrument(tracing::info_span!("tcp_server", %addr)),
+    );
 
     started_rx.await.expect("support tcp server started");
